+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix description of `Return`
+
 ### Changed
 - Handle aborting tests better in the Unity state window.
+- Simplify unnecessarily complex timeout time formatting
 
 ## [4.0.0] - 2021-03-28
 ### Changed